fix Signed-off-by: Nikolay Demchuk <nikolay.demchuk@flant.com>
