+++ e2e/tests/common_scheduler_test.go
@@ -150,7 +150,7 @@ var _ = Describe("Common Scheduler Extender", Ordered, func() {
 				cleanupE2ELVMVolumeGroups(ctx, k8sClient)
 			}
 
-			// Nested cluster teardown runs in AfterSuite (e2eCleanupSharedTestClusterAfterSuite) so Sds Node Configurator can reuse VMs.
+			// Nested cluster teardown runs in AfterSuite (e2eCleanupNestedTestClusterAfterSuite).
 		}
 	})
 
@@ -159,7 +159,7 @@ var _ = Describe("Common Scheduler Extender", Ordered, func() {
 		Expect(waitForVirtualizationModuleReadyIfNeeded(context.Background())).To(Succeed(),
 			"virtualization module should become Ready on base cluster (retry while Reconciling)")
 		testClusterResources = cluster.CreateOrConnectToTestCluster()
-		e2eSetSharedTestClusterResources(testClusterResources)
+		e2eRegisterNestedTestCluster(testClusterResources)
 	})
 
 	////////////////////////////////////
